bdevperf: change the type of io_time_in_usec into uint64_t
Purpose: To avoid overflow, if the passed value is too
big.

Reviewed-on: https://review.gerrithub.io/392579
Tested-by: SPDK Automated Test System <sys_sgsw@intel.com>

+++ test/lib/bdev/bdevperf/bdevperf.c
@@ -544,7 +544,7 @@ static void usage(char *program_name)
 }
 
 static void
-performance_dump(int io_time_in_usec)
+performance_dump(uint64_t io_time_in_usec)
 {
 	uint32_t index;
 	unsigned lcore_id;
